feat(APISIMP-THERMO-ADMIN-CONFIG): migrate ThermographyConfigAdminRest onto generic admin-config registry
Adds `ThermographyConfigDescriptor` (`@ApplicationScoped ConfigDescriptor<ThermographyConfigIO>`,
`featureName="thermography"`) and deletes the bespoke `ThermographyConfigAdminRest`.

Path change (pre-production wire break — no FE caller):
  `/v2/admin/thermography/config` → `/v2/admin/config/thermography`

No changes to `:ThermographyConfig` entity, DAO, service, IO, or migrations.
10-test `ThermographyConfigDescriptorTest` ports PATCH-logic coverage from the
deleted `ThermographyConfigAdminRestTest`.

`aidocs/16` APISIMP-THERMO-ADMIN-CONFIG → ✓ shipped 2026-06-11.
